requirements.txt
alembic==1.1.0.dev0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==19.3.0
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
ccsm==0.9.14.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dash==1.16.3
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.18.2
GDAL==3.0.4
html5lib==1.0.1
HTMLParser==0.0.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.0.1
install==1.3.4
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.4.1
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
mirakuru==2.3.0
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
msi-perkeyrgb==2.1
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.4
nest-asyncio==1.4.1
netifaces==0.10.4
notebook==6.1.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
OWSLib==0.19.1
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.1
passlib==1.7.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.11.0
pluggy==0.13.1
port-for==0.4
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.7.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyproj==2.5.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyshark==0.4.2.11
pytest==6.1.0
pytest-postgresql==2.5.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.14.0
python-geoip==1.2
python-geoip-geolite2==2015.303
pytz==2019.3
PyX==0.15
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
scapy==2.4.4
scour==0.37
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selectolax==0.2.7
Send2Trash==1.5.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
speaklater==1.4
SQLAlchemy==1.3.12
sqlparse==0.2.4
sshtunnel==0.1.4
systemd-python==234
terminado==0.9.1
testpath==0.4.4
TLPUI==0.1
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.31
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.12.1
WTForms==2.2.1
xkit==0.0.0
zipp==1.0.0
